requirements.txt
fpdf==1.7.2
numpy==2.1.1
pillow==10.4.0
